+++ lib/screens/play_room/play_room.dart
@@ -4,7 +4,7 @@ import 'package:HumanLifeGame/models/play_room/play_room.dart';
 import 'package:HumanLifeGame/models/play_room/player_action.dart';
 import 'package:HumanLifeGame/screens/play_room/announcement.dart';
 import 'package:HumanLifeGame/screens/play_room/dice_result.dart';
-import 'package:HumanLifeGame/screens/play_room/human_life.dart';
+import 'package:HumanLifeGame/screens/play_room/play_view.dart';
 import 'package:HumanLifeGame/screens/play_room/life_event_records.dart';
 import 'package:HumanLifeGame/screens/play_room/life_stages.dart';
 import 'package:HumanLifeGame/screens/play_room/player_action.dart';
